Cập nhật phần lấy dữ liệu chấm công hanet

+++ AMMS.Hanet/Data/Response/Hanet_Checkin_Data.cs
@@ -133,7 +133,7 @@
         /// <summary>
         ///  Timestamp tại thời điểm camera checkin.
         /// </summary>
-        public double? checkinTime { get; set; }
+        public long? checkinTime { get; set; }
     }
 
 }
